Changed the app.js file - it was causing a hang. Moved routes folder to routers.
